requirements.txt
streamlit
Pillow
numpy
keras
matplotlib
tensorflow